+++ packages/firebase_app_installations/firebase_app_installations/ios/firebase_app_installations/Sources/firebase_app_installations/Constants.swift
@@ -3,4 +3,4 @@
 // found in the LICENSE file.
 
 // Auto-generated file. Do not edit.
-public let versionNumber = "0.4.0+1"
+public let versionNumber = "0.4.0+3"