+++ packages/firebase_data_connect/firebase_data_connect/example/lib/main.dart
@@ -35,20 +35,20 @@ void main() async {
     await FirebaseAppCheck.instance.activate(
       // You can also use a `ReCaptchaEnterpriseProvider` provider instance as an
       // argument for `webProvider`
-      webProvider: ReCaptchaV3Provider('your-site-key'),
+      providerWeb: ReCaptchaV3Provider('your-site-key'),
       // Default provider for Android is the Play Integrity provider. You can use the "AndroidProvider" enum to choose
       // your preferred provider. Choose from:
       // 1. Debug provider
       // 2. Safety Net provider
       // 3. Play Integrity provider
-      androidProvider: AndroidProvider.debug,
+      providerAndroid: const AndroidDebugProvider(),
       // Default provider for iOS/macOS is the Device Check provider. You can use the "AppleProvider" enum to choose
       // your preferred provider. Choose from:
       // 1. Debug provider
       // 2. Device Check provider
       // 3. App Attest provider
       // 4. App Attest provider with fallback to Device Check provider (App Attest provider is only available on iOS 14.0+, macOS 14.0+)
-      appleProvider: AppleProvider.appAttest,
+      providerApple: const AppleAppAttestProvider(),
     );
   }
   if (configureEmulator) {