+++ tests/integration_test/firebase_app_check/firebase_app_check_e2e_test.dart
@@ -26,7 +26,7 @@ void main() {
         () async {
           await expectLater(
             FirebaseAppCheck.instance.activate(
-              webProvider: ReCaptchaV3Provider(
+              providerWeb: ReCaptchaV3Provider(
                 '6Lemcn0dAAAAABLkf6aiiHvpGD6x-zF3nOSDU2M8',
               ),
             ),
@@ -76,6 +76,32 @@ void main() {
           // This will fail until this is resolved: https://github.com/dart-lang/sdk/issues/52572
         },
         skip: kIsWeb,
+      );
+
+      test(
+        'debugToken on Android',
+        () async {
+          await expectLater(
+            FirebaseAppCheck.instance.activate(
+              providerAndroid: const AndroidDebugProvider(),
+            ),
+            completes,
+          );
+        },
+        skip: defaultTargetPlatform != TargetPlatform.android,
+      );
+
+      test(
+        'debugToken on iOS',
+        () async {
+          await expectLater(
+            FirebaseAppCheck.instance.activate(
+              providerApple: const AppleDebugProvider(),
+            ),
+            completes,
+          );
+        },
+        skip: defaultTargetPlatform != TargetPlatform.iOS,
       );
     },
   );