+++ packages/_flutterfire_internals/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.3.63
+
+ - Update a dependency to the latest release.
+
+## 1.3.62
+
+ - Update a dependency to the latest release.
+
 ## 1.3.61
 
  - Update a dependency to the latest release.