+++ packages/firebase_analytics/firebase_analytics_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.6.0+3
+
+ - Update a dependency to the latest release.
+
+## 0.6.0+2
+
+ - Update a dependency to the latest release.
+
 ## 0.6.0+1
 
  - Update a dependency to the latest release.