+++ packages/firebase_app_installations/firebase_app_installations_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.4+62
+
+ - Update a dependency to the latest release.
+
+## 0.1.4+61
+
+ - Update a dependency to the latest release.
+
 ## 0.1.4+60
 
  - Update a dependency to the latest release.