+++ packages/firebase_app_installations/firebase_app_installations_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.6+20
+
+ - Update a dependency to the latest release.
+
+## 0.1.6+19
+
+ - Update a dependency to the latest release.
+
 ## 0.1.6+18
 
  - Update a dependency to the latest release.