+++ packages/firebase_auth/firebase_auth/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 6.1.1
+
+ - Update a dependency to the latest release.
+
+## 6.1.0
+
+ - **FEAT**(auth): TOTP macOS support ([#17513](https://github.com/firebase/flutterfire/issues/17513)). ([41890d62](https://github.com/firebase/flutterfire/commit/41890d62a49258df097c19fd3b90e0b5de181526))
+
 ## 6.0.2
 
  - Update a dependency to the latest release.