+++ packages/firebase_performance/firebase_performance_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.7+20
+
+ - Update a dependency to the latest release.
+
+## 0.1.7+19
+
+ - Update a dependency to the latest release.
+
 ## 0.1.7+18
 
  - Update a dependency to the latest release.